refactor: consolidate header component & remove add buttons
